[SIG-4028] Convert SubNav component to TS (#1852)
* [SIG-4028] Convert SubNav component to TS

* Removes unused exports

* Moves SubNav inner components to styled import

* Imports styled components

* Adds testid
