+++ src/signals/incident-management/containers/IncidentOverviewPage/components/SubNav/index.ts
@@ -1,5 +1,3 @@
 // SPDX-License-Identifier: MPL-2.0
 // Copyright (C) 2020 - 2021 Gemeente Amsterdam
-import SubNav from './SubNav'
-
-export default SubNav
+export { default } from './SubNav'
